test: move reusable mocks into dedicated files
