+++ .github/workflows/pr.yml
@@ -10,6 +10,7 @@ on:
       - "Directory.Build.props"
       - "Directory.Build.targets"
       - "Directory.Packages.props"
+      - "global.json"
       - "MauiMicroMvvm.sln"
       - "version.json"
       - "build.slnf"
