+++ tests/MauiMicroMvvm.Tests/Mocks/TestApp.cs
@@ -1,0 +1,7 @@
+using Microsoft.Maui.Controls;
+
+namespace MauiMicroMvvm.Tests.Mocks;
+
+internal class TestApp : Application
+{
+}
