+++ tests/MauiMicroMvvm.Tests/Mocks/TestBehavior.cs
@@ -1,0 +1,7 @@
+using Microsoft.Maui.Controls;
+
+namespace MauiMicroMvvm.Tests.Mocks;
+
+internal class TestBehavior : Behavior<TestLabel>
+{
+}
