+++ tests/MauiMicroMvvm.Tests/Mocks/TestDisposable.cs
@@ -1,0 +1,11 @@
+namespace MauiMicroMvvm.Tests.Mocks;
+
+internal class TestDisposable : IDisposable
+{
+    public bool Disposed { get; private set; }
+
+    public void Dispose()
+    {
+        Disposed = true;
+    }
+}
