+++ tests/MauiMicroMvvm.Tests/Mocks/TestPage.cs
@@ -1,0 +1,7 @@
+using Microsoft.Maui.Controls;
+
+namespace MauiMicroMvvm.Tests.Mocks;
+
+internal class TestPage : Page
+{
+}
