+++ tests/MauiMicroMvvm.Tests/Mocks/TestShell.cs
@@ -1,0 +1,7 @@
+using Microsoft.Maui.Controls;
+
+namespace MauiMicroMvvm.Tests.Mocks;
+
+internal class TestShell : Shell
+{
+}
