+++ tests/MauiMicroMvvm.Tests/Mocks/TestViewModel.cs
@@ -1,0 +1,5 @@
+namespace MauiMicroMvvm.Tests.Mocks;
+
+internal class TestViewModel
+{
+}
